api: ledger file helpers to their own module
